OSPF
The reason for creating this diagram is to get an understanding of how routers share ip routes between each other allowing for internal networks to communicate without the need for static routes on every router. This saves time and the error creation of trying to setup direct routing.

The network  depicts a small to medium infrastructure with a few departments for data sharing and storage, with the idea of future expansion with various switches connected.